Phase 1: extract features/tabs.js (tab create/switch/close + tab bar + drop zone)
Tab management moves out of main.js into features/tabs.js. The render core
(renderMarkdown) and the desktop file-watch/session layer are supplied via
configureTabs DI so this module has no import cycle with platform/desktop.js
(still in main.js this commit). DI vars are named to avoid colliding with
those globals under the eval test harness.

Pure refactor: npm run build + npm run lint + 297 Jest tests green.
